- Fixed: CopyPlayer() in p_saveg.cpp should use normal assignment, not   memcpy to copy the player structures. - Fixed compilation with MinGW again and removed most of the new warnings.
And following is the log that I forgot to paste in for the previous commit:

- Changed the memory management for FString. Instead of using a garbage
  collected heap, it now uses normal heap calls and reference counting to
  implement lazy copying. You may now use bitwise operators to move
  (but not copy!) FStrings around in memory. This means that the
  CopyForTArray template function is gone, since TArrays can now freely
  move their contents around without bothering with their specifics.
  
  There is one important caveat, however. It is not acceptable to blindly 0
  an FString's contents. This necessitated the creation of a proper
  constructor for player_s so that it can be reset without using memset. I
  did a quick scan of all memsets in the source and didn't see anything else
  with a similar problem, but it's possible I missed something.
- Fixed: Build tiles were never deallocated.
- Fixed: Using Build's palette.dat only got half the palette right.


SVN r117 (trunk)

+++ src/c_console.cpp
@@ -1971,7 +1971,7 @@ static void C_TabComplete (bool goForward)
 		}
 		else
 		{ // Find the last matching tab, then go one past it.
-			while (++TabPos < TabCommands.Size())
+			while (++TabPos < (int)TabCommands.Size())
 			{
 				if (FindDiffPoint (TabCommands[TabPos].TabName, (char *)(CmdLine + TabStart)) < TabSize)
 				{
@@ -1986,7 +1986,7 @@ static void C_TabComplete (bool goForward)
 		}
 	}
 
-	if ((goForward && ++TabPos == TabCommands.Size()) ||
+	if ((goForward && ++TabPos == (int)TabCommands.Size()) ||
 		(!goForward && --TabPos < 0))
 	{
 		TabbedLast = false;
@@ -2022,7 +2022,7 @@ static bool C_TabCompleteList ()
 	nummatches = 0;
 	maxwidth = 0;
 
-	for (i = TabPos; i < TabCommands.Size(); ++i)
+	for (i = TabPos; i < (int)TabCommands.Size(); ++i)
 	{
 		if (FindDiffPoint (TabCommands[i].TabName, (char *)(CmdLine + TabStart)) < TabSize)
 		{
